# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/transactions": {
      "get": {
        "operationId": "TransactionController_list",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "income",
                "expense"
              ]
            }
          },
          {
            "name": "source",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "manual",
                "transfer"
              ]
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "start",
            "required": false,
            "in": "query",
            "description": "Start date (inclusive, YYYY-MM-DD)",
            "schema": {
              "example": "2026-05-01",
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": false,
            "in": "query",
            "description": "End date (inclusive, YYYY-MM-DD)",
            "schema": {
              "example": "2026-05-31",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated transaction list (newest first)",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/PaginatedDataDto"
                            },
                            {
                              "properties": {
                                "items": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/TransactionResponseDto"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List transactions with optional filters and pagination",
        "tags": [
          "Transaction"
        ]
      },
      "post": {
        "operationId": "TransactionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CreateTransactionResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a manual transaction (income or expense). Expense is rejected if it would overdraw the account.",
        "tags": [
          "Transaction"
        ]
      }
    },
    "/api/transactions/{id}": {
      "get": {
        "operationId": "TransactionController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TransactionResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a single transaction owned by the current user",
        "tags": [
          "Transaction"
        ]
      },
      "put": {
        "operationId": "TransactionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a manual transaction. Transfer-sourced transactions are immutable.",
        "tags": [
          "Transaction"
        ]
      },
      "delete": {
        "operationId": "TransactionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a manual transaction and reverse its effect on the balance. Transfer-sourced transactions are immutable.",
        "tags": [
          "Transaction"
        ]
      }
    },
...
    "/api/reports/monthly": {
      "get": {
        "operationId": "ReportController_monthly",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "description": "Month (1–12)",
            "schema": {
              "minimum": 1,
              "maximum": 12,
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1970,
              "maximum": 9999,
              "example": 2026,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/MonthlyReportResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Income / expense / net savings for a specific month",
        "tags": [
          "Report"
        ]
      }
    },
    "/api/reports/category": {
      "get": {
        "operationId": "ReportController_category",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "example": "expense",
              "type": "string",
              "enum": [
                "income",
                "expense"
              ]
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 12,
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1970,
              "maximum": 9999,
              "example": 2026,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CategoryReportResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Per-category breakdown for a given transaction type. Optional month/year (both-or-neither). Sorted by total descending.",
        "tags": [
          "Report"
        ]
      }
    },
    "/api/reports/trends": {
      "get": {
        "operationId": "ReportController_trends",
        "parameters": [
          {
            "name": "months",
            "required": false,
            "in": "query",
            "description": "Number of past months to include (current month included)",
            "schema": {
              "minimum": 1,
              "maximum": 24,
              "default": 6,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TrendItemDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Income vs expense across the last N months (current month included). Default 6, max 24.",
        "tags": [
          "Report"
        ]
      }
    },
    "/api/reports/weekly": {
      "get": {
        "operationId": "ReportController_weekly",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 12,
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1970,
              "maximum": 9999,
              "example": 2026,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/WeeklyItemDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Income vs expense bucketed by calendar week (1-7, 8-14, …) within a given month. Returns 4 or 5 buckets.",
        "tags": [
          "Report"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sonsam Saving API",
    "description": "Personal Finance Tracker — REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200,
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "example": "Success",
            "description": "Human-readable message"
          },
          "data": {
            "type": "object",
            "description": "Response payload (null when there is no body)"
          }
        },
        "required": [
          "statusCode",
          "message",
          "data"
        ]
      },
...
      "PaginationDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 45
          }
        },
        "required": [
          "page",
          "limit",
          "total"
        ]
      },
      "PaginatedDataDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationDto"
          }
        },
        "required": [
          "items",
          "pagination"
        ]
      },
      "CategoryRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "7"
          },
          "name": {
            "type": "string",
            "example": "Food"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "TransactionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "amount": {
            "type": "string",
            "example": "15.50",
            "description": "Decimal serialized as string"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "expense"
          },
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "transfer"
            ],
            "example": "manual"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryRefDto"
          },
          "date": {
            "type": "string",
            "example": "2026-05-03",
            "description": "YYYY-MM-DD"
          },
          "note": {
            "type": "object",
            "example": "Lunch",
            "nullable": true
          },
          "refId": {
            "type": "object",
            "example": null,
            "nullable": true,
            "description": "Transfer ID when source = transfer, otherwise null"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-05-03T12:30:00.000Z"
          }
        },
        "required": [
          "id",
          "amount",
          "type",
          "source",
          "category",
          "date",
          "createdAt"
        ]
      },
      "CreateTransactionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "10"
          },
          "amount": {
            "type": "string",
            "example": "500.00"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "income"
          },
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "transfer"
            ],
            "example": "manual"
          },
          "newBalance": {
            "type": "string",
            "example": "750.00",
            "description": "Account balance after applying this transaction"
          }
        },
        "required": [
          "id",
          "amount",
          "type",
          "source",
          "newBalance"
        ]
      },
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 500,
            "description": "Positive amount, max 2 decimals"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "income"
          },
          "categoryId": {
            "type": "number",
            "example": 1,
            "description": "Category ID — must belong to the user and match the transaction type"
          },
          "date": {
            "type": "string",
            "example": "2026-05-01",
            "description": "YYYY-MM-DD"
          },
          "note": {
            "type": "string",
            "example": "Monthly salary",
            "maxLength": 500
          }
        },
        "required": [
          "amount",
          "type",
          "categoryId",
          "date"
        ]
      },
      "UpdateTransactionDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 520
          },
          "categoryId": {
            "type": "number",
            "example": 1
          },
          "date": {
            "type": "string",
            "example": "2026-05-01"
          },
          "note": {
            "type": "string",
            "example": "Salary + bonus",
            "maxLength": 500
          }
        }
      },
...
      "MonthlyReportResponseDto": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string",
            "example": "May 2026"
          },
          "totalIncome": {
            "type": "string",
            "example": "1000.00"
          },
          "totalExpense": {
            "type": "string",
            "example": "270.00"
          },
          "netSavings": {
            "type": "string",
            "example": "730.00",
            "description": "totalIncome − totalExpense"
          }
        },
        "required": [
          "month",
          "totalIncome",
          "totalExpense",
          "netSavings"
        ]
      },
      "CategoryBreakdownItemDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "example": "Food"
          },
          "total": {
            "type": "string",
            "example": "120.00",
            "description": "Sum of amounts in this category"
          },
          "count": {
            "type": "number",
            "example": 8,
            "description": "Number of transactions"
          },
          "percentage": {
            "type": "number",
            "example": 44.4,
            "description": "Share of grandTotal as a percentage (rounded to 1 decimal)"
          }
        },
        "required": [
          "category",
          "total",
          "count",
          "percentage"
        ]
      },
      "CategoryReportResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "expense"
          },
          "period": {
            "type": "string",
            "example": "May 2026",
            "description": "Period label, or \"All time\""
          },
          "grandTotal": {
            "type": "string",
            "example": "270.00"
          },
          "breakdown": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryBreakdownItemDto"
            }
          }
        },
        "required": [
          "type",
          "period",
          "grandTotal",
          "breakdown"
        ]
      },
      "TrendItemDto": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string",
            "example": "May 2026"
          },
          "income": {
            "type": "string",
            "example": "1000.00"
          },
          "expense": {
            "type": "string",
            "example": "270.00"
          },
          "savings": {
            "type": "string",
            "example": "730.00",
            "description": "income − expense"
          }
        },
        "required": [
          "month",
          "income",
          "expense",
          "savings"
        ]
      },
      "WeeklyItemDto": {
        "type": "object",
        "properties": {
          "week": {
            "type": "string",
            "example": "Week 1"
          },
          "startDate": {
            "type": "string",
            "example": "2026-05-01",
            "description": "Inclusive (YYYY-MM-DD)"
          },
          "endDate": {
            "type": "string",
            "example": "2026-05-07",
            "description": "Inclusive (YYYY-MM-DD)"
          },
          "income": {
            "type": "string",
            "example": "500.00"
          },
          "expense": {
            "type": "string",
            "example": "300.00"
          },
          "savings": {
            "type": "string",
            "example": "200.00",
            "description": "income − expense"
          }
        },
        "required": [
          "week",
          "startDate",
          "endDate",
          "income",
          "expense",
          "savings"
        ]
      },
...
    }
  }
}